silentpayments: implement shared secret creation
Add function for calculating either a*B or A*b. This function is used by both
the sender and the recipient (i.e. a*B == A*b).

Add functions for generating the tagged input hash, as this is used when
creating the shared secret.

In a later commit, `create_shared_secret` will be exposed in the API for
the receiver to use when scanning as a light client.