silentpayments: compute input data from tx inputs
Add a function for taking the public data from a transaction (i.e. the
public keys and smallest outpoint) and returning the summed public key
and (optionaly) the `input_hash`.

The two are returned separately or together depending on whether this
function is called during scanning or called to aggregate the data for
later scanning (either by light clients or wallet rescans).