silentpayments: add recipient light client support
Expose a shared secret routine in the public API and a output_pubkey
creation routine. These are simple wrappers around the same functions
the module is using under the hood. They are exposed for the light
client as the minimal set of operations they need to do scanning without
access to the transaction outputs.

This means the light client will need to manage their own scanning
state, so wherever possible it is preferrable to use the
`_recipient_scan_ouputs` function.